+++ apps/scan/src/app/(app)/(home)/agentic-commerce/layout.tsx
@@ -1,0 +1,9 @@
+import { DocumentationPage } from "@/components/documentation-page";
+
+export default function AgenticCommerceLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <DocumentationPage>{children}</DocumentationPage>;
+}
