+++ apps/scan/src/app/(app)/(home)/x402/layout.tsx
@@ -1,0 +1,9 @@
+import { DocumentationPage } from "@/components/documentation-page";
+
+export default function X402Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <DocumentationPage>{children}</DocumentationPage>;
+}
